청원분야 카테고리 component 화
