+++ src/style/PetitionWrite.module.css
@@ -30,37 +30,6 @@
   margin-bottom: 20px;
 }
 
-.typeButtons {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 50px;
-  width: 100%;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.typeButton {
-  color: #132d5a;
-  font-size: 15px;
-  cursor: pointer;
-  height: 50px;
-  width: calc(100% / 3);
-  background-color: #fafafa;
-  border: 1px solid #b7b7b7;
-  border-radius: 5px;
-}
-
-.typeButton + .typeButton {
-  margin-left: 15px;
-}
-
-.selectedTypeButton {
-  color: white;
-  background-color: #132d5a;
-  border: none;
-}
-
 .inputTitle {
   height: 40px;
   width: 100%;
